pyproject.toml
[project]
name = "chitrapat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "faker>=37.4.2",
    "psycopg2-binary>=2.9.10",
]
